add if in migrate

+++ database/migrations/2022_01_15_104352_remove_is_azo_and_is_alifshop_columns_to_stores_table.php
@@ -13,10 +13,12 @@ class RemoveIsAzoAndIsAlifshopColumnsToStoresTable extends Migration
      */
     public function up()
     {
-        Schema::table('stores', function (Blueprint $table) {
-            $table->dropColumn('is_azo');
-            $table->dropColumn('is_alifshop');
-        });
+        if(config('app.env') == 'production') {
+            Schema::table('stores', function (Blueprint $table) {
+                $table->dropColumn('is_azo');
+                $table->dropColumn('is_alifshop');
+            });
+        }
     }
 
     /**
